IMP: Fix conveyor stat corruption bug; provide board that exercises it.
When two items with stats are next to each other around a conveyor, the
conveyor first moves the first and updates its stats. But if the first
item is now placed where the second used to be, then when the conveyor
tries to move the second object, it will pick up the already-moved first
object's stats when moving stats.

The bug is much easier to exercise on Linux (I suspect timing differences
elsewhere), but I have constructed a board that demonstrates it on the
original ZZT v3.2 as well.