Hopefully fix pointer arithmetic for good. Add compile scripts.
Split compilation scripts into one for release and one for debug
Remove RangeCheck option from .pas file and put it in as a compile
	parameter for debug instead.

The pointer arithmetic was never really fixed, it just so happened that
if FreePascal is invoked with -gv, then seg/ofs works, but if not, it
doesn't. The actual culprit is that FreePascal expects ^string to be a
pointer to a Pascal-style string, but ZZT uses its own strings with its
own length field. Changing ^string to ^byte and then removing seg/ofs
fixes the problem. (Or seems to.)