IMP: Remove titleOnly from WorldLoad. (from master)
This makes it easy to go directly into the editor instead of having to
load the world anew if it's the first one you loaded. (Presumably,
titleOnly's purpose was to keep loading times low when starting ZZT
from a floppy.)

Furthermore, titleOnly interferes with a heuristic in the fuzz branch.
It could be fixed, but it seems easier to just remove the feature
altogether and get rid of the inconvenience.