Fix bugs involving stale statIds and moving things onto (0,0).
Also provide test case for the latter.

The first bug can happen when e.g. a star pushes a line of boulders
and crushes a lion with an earlier stat ID, and the star has the last
stat ID of them all. Then its stat ID is suddenly greater than the max
stat ID, which causes problems. Even if the star is not the last ID, the
index can get confused and move another tile. Fixed by making a push
function that returns the new ID of the element that called it.