Fix bug that led to stuff with stats not being created after a while.
The problem was that AddStat used TStat._packed_size(), which
overestimates the size of the object that's added, and would thus
compound to slew BoardLen further and further away from the true
board length.

Now we just call a serialization every time to make sure the board
length doesn't desynchronize.

The Mandelbrot renderer now works. Add also a time estimate of how long
it takes :-)