Remove more ptoc stuff from board.*, world.*, and elements.*
I want to pull in the board code to generate ZZT boards in the ZZT puzzle
generator, but I don't want to have to pull all of the Pascal-to-C stuff
along. Hence this. It's kind of rough around the edges still (thinking
particularly of the string(foo.c_str()) constructs to bridge std::string
and ptoc string in editor.cxx and game.cxx), but it does its job.